use DWORD

+++ src/Datadog.Trace.ClrProfiler.Native/clr_helpers.cpp
@@ -5,9 +5,9 @@ namespace trace {
 AssemblyInfo GetAssemblyInfo(ICorProfilerInfo3* info,
                              const AssemblyID& assembly_id) {
   std::wstring name(kNameMaxSize, 0);
-  unsigned long name_len = 0;
-  auto hr = info->GetAssemblyInfo(assembly_id, (unsigned long)(name.size()),
-                                  &name_len, name.data(), nullptr, nullptr);
+  DWORD name_len = 0;
+  auto hr = info->GetAssemblyInfo(assembly_id, (DWORD)(name.size()), &name_len,
+                                  name.data(), nullptr, nullptr);
   if (FAILED(hr) || name_len == 0) {
     return {};
   }
@@ -23,10 +23,11 @@ std::wstring GetAssemblyName(
     return L"";
   }
   std::wstring name(kNameMaxSize, 0);
-  unsigned long name_len = 0;
+  DWORD name_len = 0;
   ASSEMBLYMETADATA assembly_metadata{};
-  unsigned long assembly_flags = 0;
-  hr = assembly_import->GetAssemblyProps(current, nullptr, nullptr, nullptr, name.data(), (unsigned long)(name.size()),
+  DWORD assembly_flags = 0;
+  hr = assembly_import->GetAssemblyProps(
+      current, nullptr, nullptr, nullptr, name.data(), (DWORD)(name.size()),
       &name_len, &assembly_metadata, &assembly_flags);
   if (FAILED(hr) || name_len == 0) {
     return L"";
@@ -38,11 +39,11 @@ std::wstring GetAssemblyName(
     const ComPtr<IMetaDataAssemblyImport>& assembly_import,
     const mdAssemblyRef& assembly_ref) {
   std::wstring name(kNameMaxSize, 0);
-  unsigned long name_len = 0;
+  DWORD name_len = 0;
   ASSEMBLYMETADATA assembly_metadata{};
-  unsigned long assembly_flags = 0;
+  DWORD assembly_flags = 0;
   auto hr = assembly_import->GetAssemblyRefProps(
-      assembly_ref, nullptr, nullptr, name.data(), (unsigned long)(name.size()),
+      assembly_ref, nullptr, nullptr, name.data(), (DWORD)(name.size()),
       &name_len, &assembly_metadata, nullptr, nullptr, &assembly_flags);
   if (FAILED(hr) || name_len == 0) {
     return L"";
@@ -54,21 +55,20 @@ FunctionInfo GetFunctionInfo(const ComPtr<IMetaDataImport>& metadata_import,
                              const mdToken& token) {
   mdToken parent_token = mdTokenNil;
   std::wstring function_name(kNameMaxSize, 0);
-  unsigned long function_name_len = 0;
+  DWORD function_name_len = 0;
 
   HRESULT hr = -1;
   switch (TypeFromToken(token)) {
     case mdtMemberRef:
       hr = metadata_import->GetMemberRefProps(
           token, &parent_token, function_name.data(),
-          (unsigned long)(function_name.size()), &function_name_len, nullptr,
-          nullptr);
+          (DWORD)(function_name.size()), &function_name_len, nullptr, nullptr);
       break;
     case mdtMethodDef:
       hr = metadata_import->GetMemberProps(
           token, &parent_token, function_name.data(),
-          (unsigned long)(function_name.size()), &function_name_len, nullptr,
-          nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, nullptr);
+          (DWORD)(function_name.size()), &function_name_len, nullptr, nullptr,
+          nullptr, nullptr, nullptr, nullptr, nullptr, nullptr);
       break;
   }
   if (FAILED(hr) || function_name_len == 0) {
@@ -83,12 +83,12 @@ FunctionInfo GetFunctionInfo(const ComPtr<IMetaDataImport>& metadata_import,
 
 ModuleInfo GetModuleInfo(ICorProfilerInfo3* info, const ModuleID& module_id) {
   std::wstring module_path(260, 0);
-  unsigned long module_path_len = 0;
+  DWORD module_path_len = 0;
   LPCBYTE base_load_address;
   AssemblyID assembly_id = 0;
-  unsigned long module_flags = 0;
+  DWORD module_flags = 0;
   HRESULT hr = info->GetModuleInfo2(
-      module_id, &base_load_address, (unsigned long)(module_path.size()),
+      module_id, &base_load_address, (DWORD)(module_path.size()),
       &module_path_len, module_path.data(), &assembly_id, &module_flags);
   if (FAILED(hr) || module_path_len == 0) {
     return {};
@@ -101,27 +101,26 @@ TypeInfo GetTypeInfo(const ComPtr<IMetaDataImport>& metadata_import,
                      const mdToken& token) {
   mdToken parent_token = mdTokenNil;
   std::wstring type_name(kNameMaxSize, 0);
-  unsigned long type_name_len = 0;
+  DWORD type_name_len = 0;
 
   HRESULT hr = -1;
   switch (TypeFromToken(token)) {
     case mdtTypeDef:
       hr = metadata_import->GetTypeDefProps(token, type_name.data(),
-                                            (unsigned long)(type_name.size()),
+                                            (DWORD)(type_name.size()),
                                             &type_name_len, nullptr, nullptr);
       break;
     case mdtTypeRef:
       hr = metadata_import->GetTypeRefProps(
-          token, &parent_token, type_name.data(),
-          (unsigned long)(type_name.size()), &type_name_len);
+          token, &parent_token, type_name.data(), (DWORD)(type_name.size()),
+          &type_name_len);
       break;
     case mdtTypeSpec:
       // do we need to handle this case?
       break;
     case mdtModuleRef:
-      metadata_import->GetModuleRefProps(token, type_name.data(),
-                                         (unsigned long)(type_name.size()),
-                                         &type_name_len);
+      metadata_import->GetModuleRefProps(
+          token, type_name.data(), (DWORD)(type_name.size()), &type_name_len);
       break;
     case mdtMemberRef:
       return GetFunctionInfo(metadata_import, token).type;
